requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
colorzero==2.0
docopt==0.6.2
gpiozero==1.6.2
idna==3.4
netifaces==0.11.0
Pillow==9.3.0
pkg_resources==0.0.0
requests==2.28.1
RPi.GPIO==0.7.1
tinydb==4.7.0
urllib3==1.26.13
yarg==0.1.9
